Improve download experience pt 1

+++ frameioclient/download.py
@@ -10,9 +10,14 @@ class FrameioDownloader(object):
   def download(self):
     original_filename = self.asset['name']
     final_destination = os.path.join(self.download_folder, original_filename)
-    
-    url = self.asset['original']
-    r = requests.get(url)
-    
-    open(final_destination, 'wb').write(r.content)
+
+    print(f"Final destiation: {final_destination}")
+
+    if os.path.isfile(final_destination):
+      return final_destination
+    else:
+      url = self.asset['original']
+      r = requests.get(url)
+      open(final_destination, 'wb').write(r.content)
+      return final_destination
